hc-vitaminder-py/requirements.txt
PyQt5==5.13.1
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.3
pyserial==3.4
